+++ abquant/gateway/binancec/__init__.py
@@ -34,7 +34,8 @@ STATUS_BINANCEC2AB: Dict[str, Status] = {
 ORDERTYPE_AB2BINANCEC: Dict[OrderType, Tuple[str, str]] = {
     OrderType.LIMIT: ("LIMIT", "GTC"),
     OrderType.MARKET: ("MARKET", "GTC"),
-    OrderType.POSTONLYLIMIT: ("LIMIT", "GTX")
+    OrderType.POSTONLYLIMIT: ("LIMIT", "GTX"),
+    OrderType.STOP_MARKET: ("STOP_MARKET", "GTC")
      # TODO
     # fak: ("LIMIT", "IOC"),
     # fok: ("LIMIT", "FOK"),